chore: finish hygiene cleanup (mojibake, CI, one-type-per-file)
- Remove remaining mojibake from GITHUB_ACTIONS_TROUBLESHOOTING.md,
  release.ps1 and validate-packages.ps1; drop references to the
  non-existent test-packaging.yml workflow
- check_one_type_per_file.py: translate French comments to English and
  extend the detection regex to also catch interface/record/delegate
  (it previously matched only class/struct/enum, silently missing real
  violations such as ISeriesInterfaces.cs)
- Split multi-type files so each file holds exactly one type:
  ISeriesInterfaces.cs -> ISeries/ILineSeries/IScatterSeries/IBarSeries;
  DateTimeUnit, ReflectionPropertyPathResolver and ResamplingStats moved
  to dedicated files. Repo-wide one-type-per-file check now passes.
- CI: replace deprecated actions/create-release@v1 with
  softprops/action-gh-release@v2; restrict nuget-publish.yml to stable
  tags (v[0-9]+.[0-9]+.[0-9]+) so prerelease tags no longer trigger both
  the publish and prerelease workflows
